+++ src/core/prelude/structures/sync-promise/index.ts
@@ -14,7 +14,6 @@ import {
 
 	ResolveHandler,
 	RejectHandler,
-	FinallyHandler,
 
 	ConstrRejectHandler,
 	ConstrFulfillHandler
@@ -28,13 +27,21 @@ export * from 'core/prelude/structures/sync-promise/interface';
  */
 export default class SyncPromise<T = unknown> implements Promise<T> {
 	/**
-	 * Creates a new resolved promise for the specified value
+	 * Returns a SyncPromise object that is resolved with a given value.
+	 *
+	 * If the value is a promise, that promise is returned; if the value is a thenable (i.e., has a "then" method),
+	 * the returned promise will "follow" that thenable, adopting its eventual state; otherwise,
+	 * the returned promise will be fulfilled with the value.
+	 *
+	 * This function flattens nested layers of promise-like objects
+	 * (e.g., a promise that resolves to a promise that resolves to something) into a single layer.
+	 *
 	 * @param value
 	 */
 	static resolve<T = unknown>(value: Value<T>): SyncPromise<T>;
 
 	/**
-	 * Creates a new resolved promise
+	 * Returns a new resolved SyncPromise object with an undefined value
 	 */
 	static resolve(): SyncPromise<void>;
 	static resolve<T = unknown>(value?: Value<T>): SyncPromise<T> {
@@ -49,7 +56,7 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Creates a new rejected promise with the specified reason
+	 * Returns a SyncPromise object that is rejected with a given reason
 	 * @param [reason]
 	 */
 	static reject<T = never>(reason?: unknown): SyncPromise<T> {
@@ -58,8 +65,10 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Creates a promise that is resolved with an array of results when all the provided promises
-	 * are resolved, or rejected when any promise is rejected
+	 * Takes an iterable of promises and returns a single SyncPromise that resolves to an array of the results
+	 * of the input promises. This returned promise will resolve when all of the input's promises have been resolved or
+	 * if the input iterable contains no promises. It rejects immediately upon any of the input promises rejecting or
+	 * non-promises throwing an error and will reject with this first rejection message/error.
 	 *
 	 * @param values
 	 */
@@ -92,28 +101,30 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	): SyncPromise<Array<T extends Iterable<Value<infer V>> ? V : unknown>> {
 		return new SyncPromise((resolve, reject) => {
 			const
-				promises = <SyncPromise[]>[],
-				resolved = <unknown[]>[];
+				promises = <SyncPromise[]>[];
 
 			for (const el of values) {
 				promises.push(SyncPromise.resolve(el));
 			}
 
 			if (promises.length === 0) {
-				resolve(resolved);
+				resolve([]);
 				return;
 			}
 
+			const
+				results = new Array(promises.length);
+
 			let
-				counter = 0;
+				done = 0;
 
 			for (let i = 0; i < promises.length; i++) {
 				const onFulfilled = (val) => {
-					counter++;
-					resolved[i] = val;
+					done++;
+					results[i] = val;
 
-					if (counter === promises.length) {
-						resolve(resolved);
+					if (done === promises.length) {
+						resolve(results);
 					}
 				};
 
@@ -123,8 +134,14 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Creates a promise that is resolved with an array of results when all the provided promises
-	 * are resolved or rejected
+	 * Returns a promise that resolves after all of the given promises have either been fulfilled or rejected,
+	 * with an array of objects describing each promise's outcome.
+	 *
+	 * It is typically used when you have multiple asynchronous tasks that are not dependent on one another to
+	 * complete successfully, or you'd always like to know the result of each promise.
+	 *
+	 * In comparison, the SyncPromise returned by `SyncPromise.all()` may be more appropriate
+	 * if the tasks are dependent on each other / if you'd like to reject upon any of them reject immediately.
 	 *
 	 * @param values
 	 */
@@ -175,43 +192,45 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	): SyncPromise<Array<T extends Iterable<Value<infer V>> ? PromiseSettledResult<V> : PromiseSettledResult<unknown>>> {
 		return new SyncPromise((resolve) => {
 			const
-				promises = <SyncPromise[]>[],
-				resolved = <unknown[]>[];
+				promises = <SyncPromise[]>[];
 
 			for (const el of values) {
 				promises.push(SyncPromise.resolve(el));
 			}
 
 			if (promises.length === 0) {
-				resolve(resolved);
+				resolve([]);
 				return;
 			}
 
+			const
+				results = new Array(promises.length);
+
 			let
-				counter = 0;
+				done = 0;
 
 			for (let i = 0; i < promises.length; i++) {
 				const onFulfilled = (value) => {
-					counter++;
-					resolved[i] = {
+					done++;
+					results[i] = {
 						status: 'fulfilled',
 						value
 					};
 
-					if (counter === promises.length) {
-						resolve(resolved);
+					if (done === promises.length) {
+						resolve(results);
 					}
 				};
 
 				const onRejected = (reason) => {
-					counter++;
-					resolved[i] = {
+					done++;
+					results[i] = {
 						status: 'rejected',
 						reason
 					};
 
-					if (counter === promises.length) {
-						resolve(resolved);
+					if (done === promises.length) {
+						resolve(results);
 					}
 				};
 
@@ -221,7 +240,9 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Creates a promise that is resolved or rejected when any of the provided promises are resolved or rejected
+	 * Returns a promise that fulfills or rejects as soon as one of the promises in an iterable fulfills or rejects,
+	 * with the value or reason from that promise
+	 *
 	 * @param values
 	 */
 	static race<T extends Iterable<Value>>(
@@ -247,8 +268,10 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Creates a promise that is resolved when any of the provided promises are resolved or
-	 * rejected if the provided all promises are rejected
+	 * Takes an iterable of SyncPromise objects and, as soon as one of the promises in the iterable fulfills,
+	 * returns a single promise that resolves with the value from that promise. If no promises in the iterable fulfill
+	 * (if all of the given promises are rejected), then the returned promise is rejected with an AggregateError,
+	 * a new subclass of Error that groups together individual errors.
 	 *
 	 * @param values
 	 */
@@ -296,35 +319,29 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Value of the current promise state
+	 * Actual promise state
 	 */
 	protected state: State = State.pending;
 
 	/**
-	 * Value of the promise
+	 * Resolved promise value
 	 */
 	protected value: unknown;
 
 	/**
-	 * List of handler for the "resolve" operation
+	 * List of handlers to handle the promise fulfilling
 	 */
 	protected fulfillHandlers: ConstrFulfillHandler[] = [];
 
 	/**
-	 * List of handler for the "reject" operation
+	 * List of handlers to handle the promise rejection
 	 */
 	protected rejectHandlers: ConstrRejectHandler[] = [];
-
-	/**
-	 * List of handler for the "finally" operation
-	 */
-	protected finallyHandlers: FinallyHandler[] = [];
 
 	constructor(executor: Executor) {
 		const clear = () => {
 			this.fulfillHandlers = [];
 			this.rejectHandlers = [];
-			this.finallyHandlers = [];
 		};
 
 		const reject = (err) => {
@@ -337,10 +354,6 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 
 			for (let o = this.rejectHandlers, i = 0; i < o.length; i++) {
 				o[i](err);
-			}
-
-			for (let o = this.finallyHandlers, i = 0; i < o.length; i++) {
-				o[i]();
 			}
 
 			setImmediate(() => {
@@ -369,10 +382,6 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 				o[i](val);
 			}
 
-			for (let o = this.finallyHandlers, i = 0; i < o.length; i++) {
-				o[i]();
-			}
-
 			clear();
 		};
 
@@ -380,7 +389,8 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Attaches callbacks for the resolution and/or rejection of the promise
+	 * Attaches handlers for the promise fulfilled and/or rejected states.
+	 * The method returns a new promise that will be resolved with a value that returns from the passed handlers.
 	 *
 	 * @param [onFulfilled]
 	 * @param [onRejected]
@@ -414,8 +424,8 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 				this.call(onFulfilled ?? resolve, [val], reject, resolve);
 			};
 
-			const rejectWrapper = (val) => {
-				this.call(onRejected ?? reject, [val], reject, resolve);
+			const rejectWrapper = (err) => {
+				this.call(onRejected ?? reject, [err], reject, resolve);
 			};
 
 			this.fulfillHandlers.push(fulfillWrapper);
@@ -428,15 +438,17 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Attaches a callback for only the rejection of the promise
+	 * Attaches a handler for the promise' rejected state.
+	 * The method returns a new promise that will be resolved with a value that returns from the passed handler.
+	 *
 	 * @param [onRejected]
 	 */
 	catch(onRejected?: Nullable<RejectHandler<T>>): SyncPromise<T>;
 	catch<R>(onRejected: RejectHandler<R>): SyncPromise<R>;
 	catch(onRejected?: RejectHandler): SyncPromise {
 		return new SyncPromise((resolve, reject) => {
-			const rejectWrapper = (val) => {
-				this.call(onRejected ?? reject, [val], reject, resolve);
+			const rejectWrapper = (err) => {
+				this.call(onRejected ?? reject, [err], reject, resolve);
 			};
 
 			this.fulfillHandlers.push(resolve);
@@ -449,23 +461,56 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Attaches a callback that is invoked when the promise is settled (fulfilled or rejected).
-	 * The resolved value cannot be modified from the callback.
+	 * Attaches a common callback for the promise fulfilled and rejected states.
+	 * The method returns a new promise with the state and value from the current.
+	 * A value from the passed callback will be ignored unless it equals a rejected promise or exception.
 	 *
 	 * @param [cb]
 	 */
-	finally(cb?: Nullable<FinallyHandler>): SyncPromise<T> {
+	finally(cb?: Nullable<Function>): SyncPromise<T> {
 		return new SyncPromise((resolve, reject) => {
-			if (cb != null) {
-				this.finallyHandlers.push(cb);
-			}
+			const fulfillWrapper = () => {
+				try {
+					let
+						res = cb?.();
 
-			this.fulfillHandlers.push(resolve);
-			this.rejectHandlers.push(reject);
+					if (Object.isPromiseLike(res)) {
+						res = res.then(() => this.value);
+
+					} else {
+						res = this.value;
+					}
+
+					resolve(res);
+
+				} catch (err) {
+					reject(err);
+				}
+			};
+
+			const rejectWrapper = () => {
+				try {
+					let
+						res = cb?.();
+
+					if (Object.isPromiseLike(res)) {
+						res = res.then(() => this.value);
+						resolve(res);
+
+					} else {
+						reject(this.value);
+					}
+
+				} catch (err) {
+					reject(err);
+				}
+			};
+
+			this.fulfillHandlers.push(fulfillWrapper);
+			this.rejectHandlers.push(rejectWrapper);
 
 			if (!this.isPending) {
-				(this.state === State.fulfilled ? resolve : reject)(this.value);
-				cb?.();
+				(this.state === State.fulfilled ? fulfillWrapper : rejectWrapper)();
 			}
 		});
 	}
@@ -474,7 +519,7 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 	 * Executes a function with the specified parameters
 	 *
 	 * @param fn
-	 * @param args - arguments of the function
+	 * @param args - arguments for the function
 	 * @param [onError] - error handler
 	 * @param [onValue] - success handler
 	 */
@@ -490,13 +535,13 @@ export default class SyncPromise<T = unknown> implements Promise<T> {
 
 		try {
 			const
-				val = fn?.(...args);
+				res = fn?.(...args);
 
-			if (Object.isPromiseLike(val)) {
-				(<PromiseLike<V>>val).then(resolve, reject);
+			if (Object.isPromiseLike(res)) {
+				(<PromiseLike<V>>res).then(resolve, reject);
 
 			} else {
-				resolve(val);
+				resolve(res);
 			}
 
 		} catch (err) {