+++ src/core/promise/abortable/index.ts
@@ -22,7 +22,6 @@ import {
 
 	ResolveHandler,
 	RejectHandler,
-	FinallyHandler,
 
 	ConstrFulfillHandler,
 	ConstrRejectHandler
@@ -56,9 +55,16 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Creates a new resolved AbortablePromise promise for the specified value.
+	 * Returns a AbortablePromise object that is resolved with a given value.
 	 * If the resolved value is a function, it will be invoked.
 	 * The result of the invoking will be provided as a value of the promise.
+	 *
+	 * If the value is a promise, that promise is returned; if the value is a thenable (i.e., has a "then" method),
+	 * the returned promise will "follow" that thenable, adopting its eventual state; otherwise,
+	 * the returned promise will be fulfilled with the value.
+	 *
+	 * This function flattens nested layers of promise-like objects
+	 * (e.g., a promise that resolves to a promise that resolves to something) into a single layer.
 	 *
 	 * @param value
 	 * @param [parent] - parent promise
@@ -66,7 +72,7 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	static resolveAndCall<T = unknown>(value: ExecutableValue<T>, parent?: AbortablePromise): AbortablePromise<T>;
 
 	/**
-	 * Creates a new resolved AbortablePromise promise
+	 * Returns a new resolved AbortablePromise object with an undefined value
 	 */
 	static resolveAndCall(): AbortablePromise<void>;
 	static resolveAndCall<T = unknown>(value?: ExecutableValue<T>, parent?: AbortablePromise): AbortablePromise<T> {
@@ -83,7 +89,14 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Creates a new resolved AbortablePromise promise for the specified value
+	 * Returns a AbortablePromise object that is resolved with a given value.
+	 *
+	 * If the value is a promise, that promise is returned; if the value is a thenable (i.e., has a "then" method),
+	 * the returned promise will "follow" that thenable, adopting its eventual state; otherwise,
+	 * the returned promise will be fulfilled with the value.
+	 *
+	 * This function flattens nested layers of promise-like objects
+	 * (e.g., a promise that resolves to a promise that resolves to something) into a single layer.
 	 *
 	 * @param value
 	 * @param [parent] - parent promise
@@ -91,7 +104,7 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	static resolve<T = unknown>(value: Value<T>, parent?: AbortablePromise): AbortablePromise<T>;
 
 	/**
-	 * Creates a new resolved AbortablePromise promise
+	 * Returns a new resolved AbortablePromise object with an undefined value
 	 */
 	static resolve(): AbortablePromise<void>;
 	static resolve<T = unknown>(value?: Value<T>, parent?: AbortablePromise): AbortablePromise<T> {
@@ -107,7 +120,7 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Creates a new rejected AbortablePromise promise for the specified reason
+	 * Returns a AbortablePromise object that is rejected with a given reason
 	 *
 	 * @param [reason]
 	 * @param [parent] - parent promise
@@ -117,8 +130,10 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Creates a AbortablePromise promise that is resolved with an array of results when all the provided promises
-	 * are resolved, or rejected when any promise is rejected
+	 * Takes an iterable of promises and returns a single AbortablePromise that resolves to an array of the results
+	 * of the input promises. This returned promise will resolve when all of the input's promises have been resolved or
+	 * if the input iterable contains no promises. It rejects immediately upon any of the input promises rejecting or
+	 * non-promises throwing an error and will reject with this first rejection message/error.
 	 *
 	 * @param values
 	 * @param [parent] - parent promise
@@ -160,15 +175,14 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	): AbortablePromise<Array<T extends Iterable<Value<infer V>> ? V : unknown>> {
 		return new AbortablePromise((resolve, reject, onAbort) => {
 			const
-				promises = <AbortablePromise[]>[],
-				resolved = <any[]>[];
+				promises = <AbortablePromise[]>[];
 
 			for (const el of values) {
-				promises.push(AbortablePromise.resolve(el));
+				promises.push(AbortablePromise.resolve(el, parent));
 			}
 
 			if (promises.length === 0) {
-				resolve(resolved);
+				resolve([]);
 				return;
 			}
 
@@ -178,16 +192,19 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 				}
 			});
 
+			const
+				results = new Array(promises.length);
+
 			let
-				counter = 0;
+				done = 0;
 
 			for (let i = 0; i < promises.length; i++) {
 				const onFulfilled = (val) => {
-					counter++;
-					resolved[i] = val;
+					done++;
+					results[i] = val;
 
-					if (counter === promises.length) {
-						resolve(resolved);
+					if (done === promises.length) {
+						resolve(results);
 					}
 				};
 
@@ -198,8 +215,14 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Creates a promise that is resolved with an array of results when all the provided promises
-	 * are resolved or rejected
+	 * Returns a promise that resolves after all of the given promises have either been fulfilled or rejected,
+	 * with an array of objects describing each promise's outcome.
+	 *
+	 * It is typically used when you have multiple asynchronous tasks that are not dependent on one another to
+	 * complete successfully, or you'd always like to know the result of each promise.
+	 *
+	 * In comparison, the AbortablePromise returned by `AbortablePromise.all()` may be more appropriate
+	 * if the tasks are dependent on each other / if you'd like to reject upon any of them reject immediately.
 	 *
 	 * @param values
 	 * @param [parent] - parent promise
@@ -263,15 +286,14 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	> {
 		return new AbortablePromise((resolve, _, onAbort) => {
 			const
-				promises = <AbortablePromise[]>[],
-				resolved = <any[]>[];
+				promises = <AbortablePromise[]>[];
 
 			for (const el of values) {
-				promises.push(AbortablePromise.resolve(el));
+				promises.push(AbortablePromise.resolve(el, parent));
 			}
 
 			if (promises.length === 0) {
-				resolve(resolved);
+				resolve([]);
 				return;
 			}
 
@@ -281,31 +303,34 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 				}
 			});
 
+			const
+				results = new Array(promises.length);
+
 			let
-				counter = 0;
+				done = 0;
 
 			for (let i = 0; i < promises.length; i++) {
 				const onFulfilled = (value) => {
-					counter++;
-					resolved[i] = {
+					done++;
+					results[i] = {
 						status: 'fulfilled',
 						value
 					};
 
-					if (counter === promises.length) {
-						resolve(resolved);
+					if (done === promises.length) {
+						resolve(results);
 					}
 				};
 
 				const onRejected = (reason) => {
-					counter++;
-					resolved[i] = {
+					done++;
+					results[i] = {
 						status: 'rejected',
 						reason
 					};
 
-					if (counter === promises.length) {
-						resolve(resolved);
+					if (done === promises.length) {
+						resolve(results);
 					}
 				};
 
@@ -331,7 +356,7 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 				promises = <AbortablePromise[]>[];
 
 			for (const el of values) {
-				promises.push(AbortablePromise.resolve(el));
+				promises.push(AbortablePromise.resolve(el, parent));
 			}
 
 			if (promises.length === 0) {
@@ -368,7 +393,7 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 				promises = <AbortablePromise[]>[];
 
 			for (const el of values) {
-				promises.push(AbortablePromise.resolve(el));
+				promises.push(AbortablePromise.resolve(el, parent));
 			}
 
 			if (promises.length === 0) {
@@ -415,7 +440,7 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	protected pendingChildren: number = 0;
 
 	/**
-	 * Value of the current promise state
+	 * Actual promise state
 	 */
 	protected state: State = State.pending;
 
@@ -510,7 +535,8 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Attaches callbacks for the resolution and/or rejection of the promise
+	 * Attaches handlers for the promise fulfilled and/or rejected states.
+	 * The method returns a new promise that will be resolved with a value that returns from the passed handlers.
 	 *
 	 * @param [onFulfilled]
 	 * @param [onRejected]
@@ -578,7 +604,9 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Attaches a callback for only the rejection of the promise
+	 * Attaches a handler for the promise' rejected state.
+	 * The method returns a new promise that will be resolved with a value that returns from the passed handler.
+	 *
 	 * @param [onRejected]
 	 */
 	catch(onRejected?: Nullable<RejectHandler<T>>): AbortablePromise<T>;
@@ -605,12 +633,13 @@ export default class AbortablePromise<T = unknown> implements Promise<T> {
 	}
 
 	/**
-	 * Attaches a callback that is invoked when the promise is settled (fulfilled or rejected).
-	 * The resolved value cannot be modified from the callback.
+	 * Attaches a common callback for the promise fulfilled and rejected states.
+	 * The method returns a new promise with the state and value from the current.
+	 * A value from the passed callback will be ignored unless it equals a rejected promise or exception.
 	 *
 	 * @param [cb]
 	 */
-	finally(cb?: Nullable<FinallyHandler>): AbortablePromise<T> {
+	finally(cb?: Nullable<Function>): AbortablePromise<T> {
 		return new AbortablePromise((resolve, reject, onAbort) => {
 			const
 				that = this;