+++ CHANGELOG.md
@@ -11,11 +11,14 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
-## v3.59.2 (2021-09-23)
+## v3.59.2 (2021-09-27)
 
 #### :bug: Bug Fix
 
 * Fixed date formatting with optional patterns `core/prelude/date/format`
+* Now `finally` respects exceptions values from a callback:
+  * `core/promise/sync`
+  * `core/promise/abortable`
 
 #### :nail_care: Polish