+++ src/core/promise/abortable/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.59.2 (2021-09-27)
+
+#### :bug: Bug Fix
+
+* Now `finally` respects exceptions values from a callback
+
 ## v3.59.0 (2021-09-08)
 
 #### :rocket: New Feature